+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @reservoir0x/relay-kit-hooks
 
+## 1.4.9
+
+### Patch Changes
+
+- Updated dependencies [a52745f]
+  - @reservoir0x/relay-sdk@1.4.7
+
+## 1.4.8
+
+### Patch Changes
+
+- 1c68705: Fix usePrice error handling
+
+## 1.4.7
+
+### Patch Changes
+
+- Updated dependencies [69df434]
+  - @reservoir0x/relay-sdk@1.4.6
+
 ## 1.4.6
 
 ### Patch Changes