+++ packages/relay-svm-wallet-adapter/CHANGELOG.md
@@ -1,4 +1,19 @@
-# @reservoir0x/relay-solana-wallet-adapter
+# @reservoir0x/relay-svm-wallet-adapter
+
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [a52745f]
+  - @reservoir0x/relay-sdk@1.4.7
+
+## 2.0.6
+
+### Patch Changes
+
+- 69df434: Support for eclipse svm
+- Updated dependencies [69df434]
+  - @reservoir0x/relay-sdk@1.4.6
 
 ## 2.0.5