+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @reservoir0x/relay-sdk
 
+## 1.4.7
+
+### Patch Changes
+
+- a52745f: Update zero address for zero chain
+
+## 1.4.6
+
+### Patch Changes
+
+- 69df434: Support for eclipse svm
+
 ## 1.4.5
 
 ### Patch Changes